adjust test case

+++ src/tsdb/src/tsdbFS.c
@@ -102,7 +102,7 @@ static int tsdbDecodeDFileSetArray(void **originBuf, void *buf, SArray *pArray, 
   if (pSFSHeader->version == TSDB_FS_VER_0) {
     uint64_t extendedSize = pSFSHeader->len + nset * TSDB_FILE_MAX * sizeof(TSDB_FVER_TYPE);
     if (taosTSizeof(*originBuf) < extendedSize) {
-      int ptrDistance = POINTER_DISTANCE(buf, *originBuf);
+      size_t ptrDistance = POINTER_DISTANCE(buf, *originBuf);
       if (tsdbMakeRoom(originBuf, extendedSize) < 0) {
         terrno = TSDB_CODE_FS_OUT_OF_MEMORY;
         return -1;
